pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "django-payments-mollie"
description = "Mollie integration for django-payments"
readme = "README.md"
requires-python = ">=3.7"
license = { file = "LICENSE.txt" }
keywords = []
authors = [
  { name = "Four Digits", email = "info@fourdigits.nl" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "django-payments >=1.0.0",
  "mollie-api-python >=3.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
  "flit",
]
test = [
  "pytest",
  "pytest-cov",
  "pytest-django",
  "dj_database_url",
  "django-webtest",
  "responses",
]

[project.urls]
Documentation = "https://github.com/fourdigits/django-payments-mollie#readme"
Issues = "https://github.com/fourdigits/django-payments-mollie/issues"
Source = "https://github.com/fourdigits/django-payments-mollie"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "testproject.test_settings"
pythonpath = ". testproject"
addopts = "--cov=django_payments_mollie/ --cov-branch --cov-report=term --cov-report=xml"
